BUG: Fix typo (forgotten cleanup)

+++ src/rows.h
@@ -33,9 +33,9 @@ int count_rows(FILE *f, parser_config *pconfig);
 */
 char *
 read_rows(stream *s,
-        int *nrows, int num_field_types, field_type *field_types,
+        Py_ssize_t *nrows, int num_field_types, field_type *field_types,
         parser_config *pconfig, int32_t *usecols, int num_usecols,
-        int skiplines, PyObject *converters, char *data_array,
+        Py_ssize_t skiplines, PyObject *converters, char *data_array,
         int *num_cols, bool homogeneous, bool needs_init);
 
 #endif
